+++ Beat/User Interface/BeatClipView.m
@@ -1,0 +1,17 @@
+//
+//  BeatClipView.m
+//  Beat
+//
+//  Created by Lauri-Matti Parppei on 9.10.2021.
+//  Copyright © 2021 KAPITAN!. All rights reserved.
+//
+
+#import "BeatClipView.h"
+
+@implementation BeatClipView
+
+- (void)drawRect:(NSRect)dirtyRect {
+    [super drawRect:dirtyRect];
+}
+
+@end