+++ Beat/User Interface/ScrollView.m
@@ -163,7 +163,9 @@
 		CGFloat height = self.findBarView.frame.size.height;
 		_outlineButtonY.constant = _buttonDefaultY + height;
 	}
+	
 }
+
 
 #pragma mark - Tagging View